Move all the styles to VS.UI, style infopanel
