+++ src/GitHub.UI/Controls/Buttons/OcticonButton.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using NullGuard;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -7,7 +8,7 @@ using System.Windows.Controls;
 
 namespace GitHub.UI
 {
-    public class OcticonButton: Button
+    public class OcticonButton : Button
     {
         public static readonly DependencyProperty IconRotationAngleProperty = DependencyProperty.Register(
             "IconRotationAngle", typeof(double), typeof(OcticonButton),
@@ -17,6 +18,16 @@ namespace GitHub.UI
         {
             get { return (double)GetValue(IconRotationAngleProperty); }
             set { SetValue(IconRotationAngleProperty, value); }
+        }
+
+        public static DependencyProperty IconProperty =
+            OcticonPath.IconProperty.AddOwner(typeof(OcticonButton));
+
+        public Octicon Icon
+        {
+            [return: AllowNull]
+            get { return (Octicon)GetValue(OcticonPath.IconProperty); }
+            set { SetValue(OcticonPath.IconProperty, value); }
         }
     }
 }
