+++ src/GitHub.UI/Controls/Buttons/OcticonButton.xaml
@@ -1,0 +1,40 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:ui="clr-namespace:GitHub.UI">
+
+    <Style x:Key="OcticonButton" TargetType="{x:Type ui:OcticonButton}">
+        <Setter Property="Focusable" Value="True"/>
+        <Setter Property="Foreground" Value="Black"/>
+        <Setter Property="Height" Value="16"/>
+        <Setter Property="Width" Value="16"/>
+        <Setter Property="HorizontalAlignment" Value="Center"/>
+        <Setter Property="VerticalAlignment" Value="Center"/>
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="ui:OcticonButton">
+                    <Grid x:Name="content" Background="Transparent">
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition Width="Auto"/>
+                        </Grid.ColumnDefinitions>
+                        <ui:FixedAspectRatioPanel HorizontalContentAlignment="Center" VerticalContentAlignment="Center" IsHitTestVisible="False">
+                            <Viewbox>
+                                <ui:OcticonPath Height="16"
+                                                Fill="{TemplateBinding Foreground}"
+                                                Icon="{TemplateBinding Icon}"
+                                                SnapsToDevicePixels="True"/>
+                            </Viewbox>
+                        </ui:FixedAspectRatioPanel>
+                    </Grid>
+                    <ControlTemplate.Triggers>
+                        <Trigger Property="IsMouseOver" Value="True">
+                            <Setter TargetName="content" Property="Background" Value="#11FFFFFF"/>
+                        </Trigger>
+                        <Trigger Property="IsPressed" Value="True">
+                            <Setter TargetName="content" Property="Background" Value="#11FFFFFF"/>
+                        </Trigger>
+                    </ControlTemplate.Triggers>                
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+    </Style>
+
+    <Style BasedOn="{StaticResource OcticonButton}" TargetType="{x:Type ui:OcticonButton}"/>
+</ResourceDictionary>
