+++ src/GitHub.VisualStudio.UI/Styles/ActionLinkButton.xaml
@@ -1,7 +1,7 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:PresentationOptions="http://schemas.microsoft.com/winfx/2006/xaml/presentation/options"
-                    xmlns:vsui="clr-namespace:GitHub.VisualStudio.UI.Views.Controls">
+                    >
 
   <Style x:Key="ActionLinkButton" TargetType="{x:Type Button}">
     <Setter Property="Background" Value="Transparent" />
