+++ src/GitHub.VisualStudio.UI/Styles/Buttons.xaml
