+++ src/GitHub.VisualStudio.UI/Styles/GitHubActionLink.xaml
