+++ src/GitHub.VisualStudio.UI/Styles/GitHubComboBox.xaml
@@ -1,0 +1,100 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:cache="clr-namespace:GitHub.VisualStudio.UI.Helpers"
+                    xmlns:ui="clr-namespace:GitHub.UI;assembly=GitHub.UI">
+
+    <ResourceDictionary.MergedDictionaries>
+        <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.UI;component/SharedDictionary.xaml"/>
+    </ResourceDictionary.MergedDictionaries>
+
+    <Style x:Key="GitHubComboBoxBorder" TargetType="{x:Type Border}">
+        <Setter Property="Margin" Value="10"/>
+        <Setter Property="BorderThickness" Value="1"/>
+        <Setter Property="BorderBrush" Value="{DynamicResource GitHubComboBoxBorderBrush}"/>
+        <Setter Property="Effect">
+            <Setter.Value>
+                <DropShadowEffect BlurRadius="5"
+                                  Direction="315"
+                                  Opacity="0.25"
+                                  ShadowDepth="5"/>
+            </Setter.Value>
+        </Setter>
+    </Style>
+
+    <Style x:Key="GitHubComboBoxGridContainer" TargetType="{x:Type Grid}">
+        <Setter Property="Background" Value="{DynamicResource GitHubComboBoxBackgroundBrush}"/>
+
+        <Style.Resources>
+            <Style BasedOn="{StaticResource {x:Type TextBlock}}" TargetType="{x:Type TextBlock}">
+                <Setter Property="Foreground" Value="{DynamicResource GitHubVsToolWindowText}"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type ui:OcticonImage}}" TargetType="{x:Type ui:OcticonImage}">
+                <Setter Property="Foreground" Value="{DynamicResource GitHubVsToolWindowText}"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type Separator}}" TargetType="{x:Type Separator}">
+                <Setter Property="Background" Value="{DynamicResource GitHubComboBoxBorderBrush}"/>
+                <Setter Property="Margin" Value="0"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type ui:FilterTextBox}}" TargetType="{x:Type ui:FilterTextBox}">
+                <Setter Property="BorderBrush" Value="{DynamicResource GitHubComboBoxBorderBrush}"/>
+                <Setter Property="Foreground" Value="{DynamicResource GitHubVsWindowText}"/>
+                <Setter Property="Background" Value="{DynamicResource GitHubVsBrandedUIBackground}"/>
+                <Setter Property="Margin" Value="5"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type ListBox}}" TargetType="{x:Type ListBox}">
+                <Setter Property="ScrollViewer.HorizontalScrollBarVisibility" Value="Disabled"/>
+                <Setter Property="BorderThickness" Value="0"/>
+                <Setter Property="Background" Value="Transparent"/>
+                <Setter Property="MaxHeight" Value="100"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type ListBoxItem}}" TargetType="{x:Type ListBoxItem}">
+                <Setter Property="Padding" Value="3"/>
+                <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
+            </Style>
+        </Style.Resources>
+    </Style>
+
+    <Style x:Key="GitHubComboBoxDockPanelContainer" TargetType="{x:Type DockPanel}">
+        <Setter Property="Background" Value="{DynamicResource GitHubComboBoxBackgroundBrush}"/>
+
+        <Style.Resources>
+            <Style BasedOn="{StaticResource {x:Type TextBlock}}" TargetType="{x:Type TextBlock}">
+                <Setter Property="Foreground" Value="{DynamicResource GitHubVsToolWindowText}"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type ui:OcticonImage}}" TargetType="{x:Type ui:OcticonImage}">
+                <Setter Property="Foreground" Value="{DynamicResource GitHubVsToolWindowText}"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type Separator}}" TargetType="{x:Type Separator}">
+                <Setter Property="Background" Value="{DynamicResource GitHubComboBoxBorderBrush}"/>
+                <Setter Property="Margin" Value="0"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type ui:FilterTextBox}}" TargetType="{x:Type ui:FilterTextBox}">
+                <Setter Property="BorderBrush" Value="{DynamicResource GitHubComboBoxBorderBrush}"/>
+                <Setter Property="Foreground" Value="{DynamicResource GitHubVsWindowText}"/>
+                <Setter Property="Background" Value="{DynamicResource GitHubVsBrandedUIBackground}"/>
+                <Setter Property="Margin" Value="5"/>
+                <Setter Property="Height" Value="25"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type ListBox}}" TargetType="{x:Type ListBox}">
+                <Setter Property="ScrollViewer.HorizontalScrollBarVisibility" Value="Disabled"/>
+                <Setter Property="BorderThickness" Value="0"/>
+                <Setter Property="Background" Value="Transparent"/>
+                <Setter Property="MaxHeight" Value="100"/>
+            </Style>
+
+            <Style BasedOn="{StaticResource {x:Type ListBoxItem}}" TargetType="{x:Type ListBoxItem}">
+                <Setter Property="Padding" Value="3"/>
+                <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
+            </Style>
+        </Style.Resources>
+    </Style>
+</ResourceDictionary>
