+++ src/GitHub.VisualStudio.UI/Styles/GitHubTabControl.xaml
