+++ src/GitHub.VisualStudio.UI/Styles/LinkDropDown.xaml
@@ -1,10 +1,10 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
+                    xmlns:cache="clr-namespace:GitHub.VisualStudio.UI.Helpers"
                     xmlns:ui="clr-namespace:GitHub.UI;assembly=GitHub.UI">
-    
+
     <ResourceDictionary.MergedDictionaries>
-        <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.UI;component/SharedDictionary.xaml" />
+        <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.UI;component/SharedDictionary.xaml"/>
     </ResourceDictionary.MergedDictionaries>
 
     <Style x:Key="HyperLinkToggleButton" TargetType="ToggleButton">
@@ -17,27 +17,25 @@
                                 <Style.Triggers>
                                     <MultiDataTrigger>
                                         <MultiDataTrigger.Conditions>
-                                            <Condition Binding="{Binding Path=IsMouseOver, RelativeSource={RelativeSource Self}}" Value="true" />
-                                            <Condition Binding="{Binding Path=IsEnabled, RelativeSource={RelativeSource Self}}" Value="true" />
-                                            <Condition Binding="{Binding Path=IsChecked, RelativeSource={RelativeSource TemplatedParent}}" Value="false" />
+                                            <Condition Binding="{Binding Path=IsMouseOver, RelativeSource={RelativeSource Self}}" Value="true"/>
+                                            <Condition Binding="{Binding Path=IsEnabled, RelativeSource={RelativeSource Self}}" Value="true"/>
+                                            <Condition Binding="{Binding Path=IsChecked, RelativeSource={RelativeSource TemplatedParent}}" Value="false"/>
                                         </MultiDataTrigger.Conditions>
                                         <MultiDataTrigger.Setters>
-                                            <Setter Property="TextDecorations" Value="Underline" />
-                                            <Setter Property="FrameworkElement.Cursor" Value="Hand" />
+                                            <Setter Property="TextDecorations" Value="Underline"/>
+                                            <Setter Property="FrameworkElement.Cursor" Value="Hand"/>
                                         </MultiDataTrigger.Setters>
                                     </MultiDataTrigger>
                                     <DataTrigger Binding="{Binding Path=IsPressed, RelativeSource={RelativeSource TemplatedParent}}" Value="true">
-                                        <Setter Property="TextDecorations" Value="Underline" />
-                                        <Setter Property="FrameworkElement.Cursor" Value="Hand" />
+                                        <Setter Property="TextDecorations" Value="Underline"/>
+                                        <Setter Property="FrameworkElement.Cursor" Value="Hand"/>
                                     </DataTrigger>
                                 </Style.Triggers>
-                                <Setter Property="Cursor" Value="Hand" />
+                                <Setter Property="Cursor" Value="Hand"/>
                             </Style>
                         </TextBlock.Style>
-                        <Run Text="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=Content}" />
-                        <Polygon Margin="2,0,0,1"
-                                Fill="{TemplateBinding Foreground}"
-                                Points="0,0 8,0 4,4 0,0"/>
+                        <Run Text="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=Content}"/>
+                        <Polygon Margin="2,0,0,1" Fill="{TemplateBinding Foreground}" Points="0,0 8,0 4,4 0,0"/>
                     </TextBlock>
                 </ControlTemplate>
             </Setter.Value>
@@ -50,25 +48,24 @@
             <Setter.Value>
                 <ControlTemplate TargetType="ui:LinkDropDown">
                     <Grid>
-                        <ToggleButton Style="{StaticResource HyperLinkToggleButton}"
-                                      Content="{TemplateBinding LinkText}"
+                        <ToggleButton Content="{TemplateBinding LinkText}"
                                       Foreground="{TemplateBinding Foreground}"
                                       IsChecked="{Binding IsDropDownOpen, Mode=TwoWay, RelativeSource={RelativeSource TemplatedParent}}"
-                                      IsTabStop="False"/>
+                                      IsTabStop="False"
+                                      Style="{StaticResource HyperLinkToggleButton}"/>
                         <Popup Name="PART_Popup"
                                AllowsTransparency="True"
-                               IsOpen="{TemplateBinding IsDropDownOpen}" 
+                               IsOpen="{TemplateBinding IsDropDownOpen}"
                                Placement="Bottom">
                             <Popup.Resources>
                                 <Style BasedOn="{StaticResource {x:Type ComboBoxItem}}" TargetType="{x:Type ComboBoxItem}">
-                                    <Setter Property="Foreground" Value="{DynamicResource GitHubButtonForegroundBrush}" />
-                                    <Setter Property="Padding" Value="3" />
-                                    <Setter Property="HorizontalContentAlignment" Value="Stretch" />
+                                    <Setter Property="Foreground" Value="{DynamicResource GitHubButtonForegroundBrush}"/>
+                                    <Setter Property="Padding" Value="3"/>
+                                    <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
                                 </Style>
                             </Popup.Resources>
                             <Border Style="{DynamicResource GitHubComboBoxBorder}">
-                                <DockPanel Style="{DynamicResource GitHubComboBoxDockPanelContainer}"
-                                           MinWidth="100">
+                                <DockPanel MinWidth="100" Style="{DynamicResource GitHubComboBoxDockPanelContainer}">
                                     <ScrollViewer VerticalScrollBarVisibility="Auto">
                                         <ItemsPresenter/>
                                     </ScrollViewer>
