+++ src/GitHub.VisualStudio.UI/Styles/TextBlocks.xaml
