+++ src/GitHub.VisualStudio.UI/Styles/ThemeDark.xaml
@@ -4,7 +4,7 @@
   xmlns:PresentationOptions="http://schemas.microsoft.com/winfx/2006/xaml/presentation/options">
 
   <ResourceDictionary.MergedDictionaries>
-    <ResourceDictionary Source="pack://application:,,,/GitHub.VisualStudio;component/Styles/VsBrush.xaml" />
+    <ResourceDictionary Source="pack://application:,,,/GitHub.VisualStudio.UI;component/Styles/VsBrush.xaml" />
   </ResourceDictionary.MergedDictionaries>
 
   <Color x:Key="GitHubContextMenuIconColor">#424242</Color>
