+++ src/GitHub.VisualStudio/Resources/icons/mark_github.xaml
@@ -8,7 +8,7 @@
             <ui:OcticonImage.Resources>
                 <ResourceDictionary>
                     <ResourceDictionary.MergedDictionaries>
-                        <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio;component/SharedDictionary.xaml" />
+                        <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio.UI;component/SharedDictionary.xaml" />
                         <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.UI;component/SharedDictionary.xaml" />
                     </ResourceDictionary.MergedDictionaries>
                 </ResourceDictionary>
