--- src/GitHub.VisualStudio/SharedDictionary.xaml
@@ -1,67 +1,0 @@
-<ResourceDictionary 
-  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:PresentationOptions="http://schemas.microsoft.com/winfx/2006/xaml/presentation/options"
-  xmlns:ui="clr-namespace:GitHub.UI;assembly=GitHub.UI"
-  xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers">
-
-  <ResourceDictionary.MergedDictionaries>
-    <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio.UI;component/SharedDictionary.xaml" />
-    <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio;component/Styles/ActionLinkButton.xaml" />
-    <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio;component/Styles/GitHubComboBox.xaml" />
-    <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio;component/Styles/Buttons.xaml" />
-    <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio;component/Styles/GitHubTabControl.xaml" />
-    <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio;component/Styles/TextBlocks.xaml" />
-    <cache:SharedDictionaryManager Source="pack://application:,,,/GitHub.VisualStudio;component/Styles/LinkDropDown.xaml" />
-  </ResourceDictionary.MergedDictionaries>
-
-  <Style x:Key="VSStyledButton" BasedOn="{StaticResource VsButtonStyleKey}" TargetType="{x:Type Button}" />
-  <Style x:Key="VSStyledCheckBox" BasedOn="{StaticResource VsCheckBoxStyleKey}" TargetType="{x:Type CheckBox}">
-    <Setter Property="Foreground" Value="{DynamicResource GitHubVsToolWindowText}" />
-    <Setter Property="Background" Value="{DynamicResource GitHubVsToolWindowBackground}" />
-    <Style.Triggers>
-      <Trigger Property="UIElement.IsMouseOver" Value="true">
-        <Setter Property="Background" Value="{DynamicResource GitHubVsToolWindowBackground}" />
-      </Trigger>
-    </Style.Triggers>
-  </Style>
-  <Style x:Key="VSStyledComboBox" BasedOn="{StaticResource VsComboBoxStyleKey}" TargetType="{x:Type ComboBox}" />
-
-  <DrawingBrush x:Key="ConnectArrowBrush">
-    <DrawingBrush.Drawing>
-      <DrawingGroup>
-        <DrawingGroup.Children>
-          <GeometryDrawing Brush="{DynamicResource GitHubActionLinkItemBrush}" Geometry="F1 M 9,11 L 7,11 9,9 4,9 4,7 9,7 7,5 9,5 12,8 Z" />
-          <GeometryDrawing Brush="{DynamicResource GitHubActionLinkItemBrush}" Geometry="F1 M 7.9741,1.0698 C 4.1461,1.0698 1.0441,4.1728 1.0441,7.9998 1.0441,11.8268 4.1461,14.9298 7.9741,14.9298 11.8011,14.9298 14.9041,11.8268 14.9041,7.9998 14.9041,4.1728 11.8011,1.0698 7.9741,1.0698 M 7.9741,2.0598 C 11.2501,2.0598 13.9151,4.7248 13.9151,7.9998 13.9151,11.2758 11.2501,13.9408 7.9741,13.9408 4.6991,13.9408 2.0341,11.2758 2.0341,7.9998 2.0341,4.7248 4.6991,2.0598 7.9741,2.0598 " />
-        </DrawingGroup.Children>
-      </DrawingGroup>
-    </DrawingBrush.Drawing>
-  </DrawingBrush>
-
-  <Style x:Key="TitleVerticalSeparator" TargetType="{x:Type Separator}">
-    <Setter Property="Foreground" Value="#BBBBBB" />
-    <Setter Property="Width" Value="2" />
-    <Setter Property="Margin" Value="3,0,3,0" />
-    <Setter Property="Template">
-      <Setter.Value>
-        <ControlTemplate TargetType="{x:Type Separator}">
-          <Border Width="{TemplateBinding Width}"
-                  Background="{TemplateBinding Foreground}"
-                  BorderBrush="{TemplateBinding Foreground}"
-                  BorderThickness="{TemplateBinding BorderThickness}"
-                  SnapsToDevicePixels="True" />
-        </ControlTemplate>
-      </Setter.Value>
-    </Setter>
-  </Style>
-
-  <Style x:Key="VerticalSeparator" TargetType="{x:Type Separator}">
-    <Setter Property="Background" Value="{DynamicResource GitHubHeaderSeparatorBrush}" />
-    <Setter Property="Margin" Value="3,0,3,0" />
-    <Setter Property="LayoutTransform">
-      <Setter.Value>
-        <RotateTransform Angle="90" />
-      </Setter.Value>
-    </Setter>
-  </Style>
-</ResourceDictionary>
